Render lobby rows via ssi
