+++ config/tic-tac-toe/services/controller.yml
@@ -1,12 +1,12 @@
 services:
     tic-tac-toe.fragment-controller:
         class: Gaming\TicTacToe\Port\Adapter\Http\FragmentController
+        arguments: ['@tic-tac-toe.query-bus', '@gaming.usernames']
         calls: [[setContainer, ['@Psr\Container\ContainerInterface']]]
         tags: ['controller.service_arguments', 'container.service_subscriber']
 
     tic-tac-toe.lobby-controller:
         class: Gaming\TicTacToe\Port\Adapter\Http\LobbyController
-        arguments: ['@tic-tac-toe.query-bus', '@gaming.usernames']
         calls: [[setContainer, ['@Psr\Container\ContainerInterface']]]
         tags: ['controller.service_arguments', 'container.service_subscriber']
