+++ src/TicTacToe/Port/Adapter/Http/View/statistics.html.twig
@@ -1,0 +1,4 @@
+<p>
+    <span class="badge bg-blue text-blue-fg">?</span>
+    running games
+</p>
